+++ HippoPrinter/src/libslic3r/GCode.cpp
@@ -391,7 +391,7 @@ GCode::extrude(ExtrusionLoop loop, std::string description, double speed)
             Polygon polygon = loop.polygon();
             Point centroid = polygon.centroid();
             last_pos = Point(polygon.bounding_box().max.x, centroid.y);
-            last_pos.rotate(fmod((float)rand()/16.0, 2.0*PI), centroid);
+            last_pos.rotate(fmod((double)rand()/16.0, 2.0*PI), centroid);
         }
         loop.split_at(last_pos);
     }
