+++ HippoPrinter/src/libslic3r/Print.hpp
@@ -163,7 +163,7 @@ class Print
     PlaceholderParser placeholder_parser;
     // TODO: status_cb
     double total_used_filament, total_extruded_volume;
-    std::map<size_t,float> filament_stats;
+    std::map<size_t,double> filament_stats;
     PrintState<PrintStep> state;
 
     // ordered collections of extrusion paths to build skirt loops and brim
